pyproject.toml
[project]
name = "my_bot_name"
version = "0.1.0"
description = "A Telegram bot ..."
authors = [
    {name = "",email = ""}
]
license = {text = "BSD 3-Clause License"}
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "uvicorn (>=0.34.2,<0.35.0)",
    "fastapi (>=0.115.12,<0.116.0)",
    "pyyaml (>=6.0.2,<7.0.0)",
    "google-api-core (>=2.24.2,<3.0.0)",
    "google-genai (>=1.11.0,<2.0.0)",
    "python-telegram-bot (>=22.0,<23.0)",
    "google-cloud-storage (>=3.1.0,<4.0.0)",
    "google-cloud-bigquery (>=3.31.0,<4.0.0)",
    "google-cloud-bigquery-storage (>=2.31.0,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
notebook = "^7.4.0"
ipywidgets = "^8.1.6"
tqdm = "^4.67.1"

